Fierce Deity fork: FD transformation mask + boot-time fd.o2r generator
Adds a fully-playable Fierce Deity transformation mask to Ship of Harkinian,
re-created from Aegiker's OoT Fierce Deity romhack and made Majora's-Mask-accurate
via the MM decompilation and 2 Ship 2 Harkinian.

- Boot-time fd.o2r generator: derives all Majora's Mask content from the user's
  own MM ROM at first launch (mirrors how SoH generates oot.o2r), so no Nintendo
  ROM assets are shipped in this repository.
- ZAPD Majora's Mask extraction support + O2rArchive filtering/patching (FdO2rGen).
- Vendored ZAPDTR / OTRExporter / libultraship (de-submoduled) so the fork clones
  and builds self-contained with the FD extraction fixes included.
- Bundled assets are only Aegiker's authored transformation audio and fork-authored
  geometry; everything sourced from Majora's Mask is generated from your ROM.

+++ soh/soh/Enhancements/Fishing.cpp
@@ -9,7 +9,7 @@ f32 Fishing_GetMinimumRequiredScore();
 
 void BuildFishingMessage(uint16_t* textId, bool* loadFromMessageTable) {
     if (gSaveContext.minigameScore == 0) {
-        gSaveContext.minigameScore = static_cast<u16>(Fishing_GetMinimumRequiredScore());
+        gSaveContext.minigameScore = Fishing_GetMinimumRequiredScore();
     }
 }
